flask-server/requirements.txt
Flask[async]
asyncio
openai
gunicorn
python-dotenv
flask-cors
python-dotenv